pyproject.toml
[project]
name = "flowconnect-recon-mvp"
version = "0.1.0"
description = "Multi-tenant invoice reconciliation API (FastAPI + Strawberry + SQLAlchemy) - demo challenge"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.112",
  "uvicorn>=0.30",
  "sqlalchemy>=2.0",
  "pydantic>=2.7",
  "strawberry-graphql[fastapi]>=0.239",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
